move string functions out of git-compat-util
We have several inline string helpers (skip_prefix, etc)
defined in git-compat-util.h. This is convenient, as it
makes them available everywhere, but this is probably not
the best place for them. Besides cluttering a file which is
otherwise mostly about compatibility issues, defining inline
functions in git-compat-util.h restricts what they can do.

For instance, if we wanted to use tolower() in one of the
functions, we run into an ordering problem with our
redefinitions of tolower(), which are done later in
git-compat-util.h.

This patch moves them to their own header file, strutil.h.
However, it also includes them in strbuf.h (which needs them
for its inline strbuf_strip_suffix). This also puts them
indirectly into cache.h, which means they are effectively
available everywhere (just as they were before).
